+++ source/common/utils/environmentCheckers.ts
@@ -1,6 +1,7 @@
 import { upperFirst } from 'lodash';
 import {
   ALONZO_PURPLE,
+  MARLOWE_PIONEERS,
   SHELLEY_QA,
   DEVELOPMENT,
   LINUX,
@@ -30,6 +31,10 @@ export const evaluateNetwork = (network: string | null | undefined) => {
     currentNetwork = ALONZO_PURPLE;
   }
 
+  if (network === 'marlowe-pioneers') {
+    currentNetwork = MARLOWE_PIONEERS;
+  }
+
   return currentNetwork;
 };
 export const getBuildLabel = (
@@ -54,6 +59,8 @@ export const checkIsMainnet = (network: string) => network === MAINNET;
 export const checkIsTestnet = (network: string) => network === TESTNET;
 export const checkIsAlonzoPurple = (network: string) =>
   network === ALONZO_PURPLE;
+export const checkIsMarlowePioneers = (network: string) =>
+  network === MARLOWE_PIONEERS;
 export const checkIsShelleyQA = (network: string) => network === SHELLEY_QA;
 export const checkIsStaging = (network: string) => network === STAGING;
 export const checkIsSelfnode = (network: string) => network === SELFNODE;